+++ opentelemetry-sdk/CHANGELOG.md
@@ -13,6 +13,10 @@
   is now the API crate. [#1226](https://github.com/open-telemetry/opentelemetry-rust/pull/1226)
 - Add in memory span exporter [#1216](https://github.com/open-telemetry/opentelemetry-rust/pull/1216)
 
+### Removed
+
+- Remove context from Metric force_flush [#1245](https://github.com/open-telemetry/opentelemetry-rust/pull/1245)
+
 ## v0.20.0
 
 ### Added